requirements.txt
Django==4.2
djangorestframework
markdown
django-filter
psycopg2-binary==2.9.7  # for dev production use psycopg2
debugpy
flake8
boto3
gunicorn
python-decouple
confluent-kafka